Add files for work
So far the script works and solves the problem, but I'm not satisfied. Let me explain.

The deployment is successful and all my files are working. However, during deployment, the phpMyAdmin service remains in an orange state, and after a certain amount of time, thanks to or because of this part I added to keep it continuously active:

livenessProbe:
httpGet:
path: /index.php
port: 80
initialDelaySeconds: 30
periodSeconds: 10
timeoutSeconds: 5
failureThreshold: 3

It becomes the top priority and turns green, but this causes my Apache server to become inactive. From what I understand, because I've encountered this issue in a previous project, Kubernetes doesn't like it when a service remains inactive for too long and puts it in a "pause" state. I'll try to improve my code so that all three services run simultaneously.

# Next steps:

- Try to put a page on Apache
- Fix the PHP that fails (why not add credentials to log in?)
- Make read me for test :)